+++ src/main/java/study/lotto/domain/LottoNumber.java
@@ -1,4 +1,4 @@
-package study.lotto.model;
+package study.lotto.domain;
 
 import java.util.*;
 import java.util.stream.Collectors;
