+++ src/main/java/study/lotto/domain/Lottos.java
@@ -1,9 +1,6 @@
-package study.lotto.model;
+package study.lotto.domain;
 
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
