+++ src/test/java/study/lotto/PrizeTest.java
@@ -4,10 +4,10 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import study.lotto.model.Lotto;
-import study.lotto.model.LottoNumber;
-import study.lotto.model.LottoRank;
-import study.lotto.model.WinningLottoInfo;
+import study.lotto.domain.Lotto;
+import study.lotto.domain.LottoNumber;
+import study.lotto.domain.LottoRank;
+import study.lotto.domain.WinningLottoInfo;
 
 import java.util.List;
 import java.util.stream.Collectors;
